Pin to tox<4
The release of tox 4.0.0 to PyPI has broken all our tox commands, for
example:

```terminal
$ make checkformatting
...
ImportError: cannot import name 'hookimpl' from 'tox'
```

In some projects I've seen this error message instead:

```
ModuleNotFoundError: No module named 'py'
```

The reason for the breakage is that tox 3.x installs the new tox 4.0.0
in the `.tox/.tox` venv. Breaking changes in tox 4.0.0 then cause
problems. In particular tox 4.0.0 breaks compatibility with all tox
plugins. Fix this by telling it to install `tox<4`.

At the same time also fix `make_python` to install `tox<4`. This
controls the version of the "outer" tox: the instance of tox that `make`
actually runs when we have commands like `pyenv exec tox` in `Makefile`.
This is just for good measure, I'm not sure that pinning the outer
version of tox is strictly necessary: if tox 4.0.0 got installed as the
outer version it would see the `tox<4` requirement in our `tox.ini` and
would install `tox<4` in the `.tox/.tox` venv and run that.

And the same in `ci.yml` too.